refactor: Modularize common types and common data
