+++ components/Resume/components/Careers/Careers.tsx
@@ -1,6 +1,8 @@
 /* eslint-disable no-unused-vars */
 import Image from 'next/image';
 
+import { CareerInputName } from '@/types/Resume';
+
 import CareerContainer from '../../containers/CareerContainer';
 
 type Props = {
@@ -20,14 +22,7 @@ type Props = {
     value,
   }: {
     id: number;
-    name:
-      | 'title'
-      | 'jobDetail'
-      | 'company'
-      | 'startDate'
-      | 'endDate'
-      | 'region'
-      | 'description';
+    name: CareerInputName;
     value: string;
   }) => void;
   onClickDeleteCareer: (id: number) => void;
