+++ fixtures/careers.ts
@@ -1,0 +1,14 @@
+const CAREERS = [
+  {
+    id: 0,
+    title: '',
+    jobDetail: '',
+    company: '',
+    startDate: '',
+    endDate: '',
+    region: '',
+    description: '',
+  },
+];
+
+export default CAREERS;
